+++ Server/src/main/java/com/dbs/dbs/models/units/Footman.java
@@ -5,7 +5,7 @@ package com.dbs.dbs.models.units;
  */
 public class Footman extends Unit{
 
-    public Footman(Integer posX, Integer posY) {
+    public Footman(Double posX, Double posY) {
         super("Footman", 3.0, 6.0, 7.0, 80.0, 50, posX, posY);
     }
 }
